refactor: maintain uniformity in file names
